photo/scripts/requirements.txt
Pillow==5.2.0
pyperclip==1.6.4
